+++ src/drivers/warmers/BaseCacheWarmer.php
@@ -73,10 +73,11 @@ abstract class BaseCacheWarmer extends SavableComponent implements CacheWarmerIn
      */
     public function warmAll(int $delay = null, callable $setProgressHandler = null)
     {
-        $siteUris = array_merge(
-            SiteUriHelper::getAllSiteUris(true),
-            Blitz::$plugin->settings->customSiteUris
-        );
+        $siteUris = SiteUriHelper::getAllSiteUris(true);
+
+        if (is_array(Blitz::$plugin->settings->customSiteUris)) {
+            $siteUris = array_merge($siteUris, Blitz::$plugin->settings->customSiteUris);
+        }
 
         $this->warmUris($siteUris, $delay, $setProgressHandler);
     }